pyproject.toml
[tool.poetry]
name = "exemplo"
version = "0.1.0"
description = ""
authors = ["Felipe Bastos Nunes <felipebastos@ifce.edu.br>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.5"
djangorestframework = "^3.13.1"
Markdown = "^3.3.7"
django-filter = "^21.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
